requirement.txt
streamlit==1.43.2
azure-ai-inference